backend/requirements.txt
flask
flask-cors
bcrypt
pyjwt
gunicorn
python-dotenv